Test pour les controllers PermissionController et ProjectController on modifie le traitement des exceptions dans quelques parts, mail n'est pas complet encore. 37 test

+++ API/src/main/java/Service/ProjectServiceImpl.java
@@ -3,6 +3,8 @@ package Service;
 import DAO.ProjectDAO;
 import DAO.ProjectDAOImpl;
 import Model.Project;
+import Model.User;
+import Model.UserGrant;
 import Util.DataException;
 
 import java.util.List;
@@ -13,7 +15,9 @@ import java.util.logging.Logger;
  * Created by amaia.nazabal on 10/21/16.
  */
 public class ProjectServiceImpl implements ProjectService {
+    UserGrantService userGrantService = new UserGrantServiceImpl();
     ProjectDAO projectDAO;
+
     private static final Logger LOGGER = Logger.getLogger( ProjectServiceImpl.class.getName() );
     public ProjectServiceImpl(){
         projectDAO = new ProjectDAOImpl();
@@ -29,13 +33,7 @@ public class ProjectServiceImpl implements ProjectService {
     }
 
     public Project getEntityById(Long id) throws DataException{
-        try {
-            return projectDAO.getEntityById(id);
-        } catch (Exception e) {
-            LOGGER.log( Level.FINE, e.toString(), e);
-        }
-
-        return null;
+        return projectDAO.getEntityById(id);
     }
 
     public List getEntityList() throws DataException{
@@ -47,14 +45,28 @@ public class ProjectServiceImpl implements ProjectService {
         return null;
     }
 
-    public boolean deleteEntity(Long id) throws DataException{
+    public boolean deleteEntity(Long idProject, Long idUser) throws DataException{
+        boolean result;
         try {
-            Project project = getEntityById(id);
-            return projectDAO.deleteEntity(project);
+            User admin = userGrantService.getAdminByEntity(idProject);
+            if (admin.getId().equals(idUser)) {
+                /* On enleve le permis avec l'admin */
+                result = userGrantService.deleteEntity(idUser, idProject, UserGrant.Permis.Admin);
+                if (result) {
+                    Project project = getEntityById(idProject);
+                    result = projectDAO.deleteEntity(project);
+                }else {
+                    throw new DataException("Error removing permission");
+                }
+            } else {
+                throw new DataException("Only the admin has permissions for remove the project");
+            }
+
+            return result;
         } catch (Exception e) {
             LOGGER.log( Level.FINE, e.toString(), e);
+            throw new DataException(e.getMessage());
         }
-        return false;
     }
 
 }
